+++ frontend/src/app/auth/auth.guard.ts
@@ -1,7 +1,7 @@
 import { inject } from "@angular/core";
 import { CanActivateFn, Router } from "@angular/router";
 import { LoginService } from "../services/login/login.service";
-import { TipoUsuario } from "../shared/models/enums/tipo-usuario.enum";
+import { TipoUsuario } from "../shared/models";
 
 export const authGuard: CanActivateFn = (route, state) => {
   const loginService = inject(LoginService);
