+++ frontend/src/app/pages/auth/login/login.component.ts
@@ -3,7 +3,7 @@ import { FormsModule, NgForm } from "@angular/forms";
 import { LoginService } from "../../../services/login/login.service";
 import { Router, RouterModule } from "@angular/router";
 import { CommonModule } from "@angular/common";
-import { TipoUsuario } from "../../../shared/models/enums/tipo-usuario.enum";
+import { TipoUsuario } from "../../../shared/models";
 
 @Component({
   selector: "tela-login",
