+++ frontend/src/app/pages/cliente/solicitacao/inserir-solicitacao/inserir-solicitacao.component.ts
@@ -1,13 +1,11 @@
 import { Component, ViewChild } from "@angular/core";
 import { FormsModule, NgForm } from "@angular/forms";
-import { Solicitacao } from "../../../../shared/models/solicitacao.model";
-import { Equipamento } from "../../../../shared/models/equipamento.model";
+import { Solicitacao,Equipamento,EstadosSolicitacao } from "../../../../shared/models";
 import { SolicitacaoService } from "../../../../services/solicitacao.service";
 import { EquipamentoService } from "../../../../services/equipamento.service";
 import { LoginService } from "../../../../services/login/login.service";
 import { Router, RouterModule } from "@angular/router";
 import { CommonModule } from "@angular/common";
-import { EstadosSolicitacao } from "../../../../shared/models/enums/estados-solicitacao";
 import { DataUtils } from "../../../../shared/utils/data-utils";
 
 @Component({
