+++ frontend/src/app/pages/empregado/equipamento/editar-equipamento/editar-equipamento.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, ViewChild } from "@angular/core";
 import { FormsModule, NgForm } from "@angular/forms";
-import { Equipamento } from "../../../../shared/models/equipamento.model";
+import { Equipamento } from "../../../../shared/models";
 import { ActivatedRoute, Router, RouterModule } from "@angular/router";
 import { EquipamentoService } from "../../../../services/equipamento.service";
 import { CommonModule } from "@angular/common";
