+++ frontend/src/app/pages/empregado/funcionario/inserir-funcionario/inserir-funcionario.component.ts
@@ -1,6 +1,6 @@
 import { Component, ViewChild } from "@angular/core";
 import { FormsModule, NgForm } from "@angular/forms";
-import { Funcionario } from "../../../../shared/models/funcionario.model";
+import { Funcionario } from "../../../../shared/models";
 import { FuncionarioService } from "../../../../services/funcionario.service";
 import { Router, RouterModule } from "@angular/router";
 import { CommonModule } from "@angular/common";
