+++ frontend/src/app/pages/empregado/funcionario/listar-funcionario/listar-funcionario.component.ts
@@ -1,6 +1,6 @@
 import { Component } from "@angular/core";
 import { FuncionarioService } from "../../../../services/funcionario.service";
-import { Funcionario } from "../../../../shared/models/funcionario.model";
+import { Funcionario } from "../../../../shared/models";
 import { CommonModule } from "@angular/common";
 import { RouterModule } from "@angular/router";
 import { LoginService } from "../../../../services/login/login.service";
