+++ frontend/src/app/pages/empregado/solicitacao/listar-atribuicao/listar-atribuicao.component.ts
@@ -3,11 +3,9 @@ import { SolicitacaoService } from "../../../../services/solicitacao.service";
 import { ClienteService } from "../../../../services/cliente.service";
 import { LoginService } from "../../../../services/login/login.service";
 import { FuncionarioService } from "../../../../services/funcionario.service";
-import { Solicitacao } from "../../../../shared/models/solicitacao.model";
-import { Cliente } from "../../../../shared/models/cliente.model";
+import { Cliente, Solicitacao,EstadosSolicitacao } from "../../../../shared/models";
 import { CommonModule } from "@angular/common";
 import { RouterModule } from "@angular/router";
-import { EstadosSolicitacao } from "../../../../shared/models/enums/estados-solicitacao";
 import { EstadoAmigavelPipe } from "../../../../shared/pipes/estado-amigavel.pipe";
 import { EstadoCorPipe } from "../../../../shared/pipes/estado-cor.pipe";
 import { HistoricoUtils } from "../../../../shared/utils/historico-utils";
