+++ frontend/src/app/services/funcionario.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Funcionario } from "../shared/models/funcionario.model";
+import { Funcionario } from "../shared/models";
 import { ServiceCrudBase } from "./service-crud-base/service-crud-base";
 
 const LS_CHAVE = "funcionarios";
