+++ frontend/src/app/services/solicitacao.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Solicitacao } from "../shared/models/solicitacao.model";
+import { Solicitacao } from "../shared/models";
 import { ServiceCrudBase } from "./service-crud-base/service-crud-base";
 
 const LS_CHAVE = "solicitacoes";
