+++ frontend/src/app/shared/models/enums/index.ts
@@ -1,0 +1,2 @@
+export * from './estados-solicitacao';
+export * from './tipo-usuario.enum';
