+++ frontend/src/app/shared/models/index.ts
@@ -1,0 +1,9 @@
+export * from './cliente.model';
+export * from './endereco';
+export * from './entidade-base.model';
+export * from './equipamento.model';
+export * from './funcionario.model';
+export * from './orcamento.model';
+export * from './sessao.model';
+export * from './solicitacao.model';
+export * from './enums'
